@@ -14,7 +14,7 @@
         }
     ],
     "require": {
-        "php": ">=5.6.4"
+        "php": ">=7.0"
     },
     "autoload": {
         "psr-4": {
@@ -23,7 +23,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "5.4-dev"
+            "dev-master": "5.5-dev"
         }
     },
     "config": {
