@@ -7,23 +7,19 @@ interface MailQueue
     /**
      * Queue a new e-mail message for sending.
      *
-     * @param  string|array  $view
-     * @param  array   $data
-     * @param  \Closure|string  $callback
+     * @param  string|array|MailableContract  $view
      * @param  string  $queue
      * @return mixed
      */
-    public function queue($view, array $data, $callback, $queue = null);
+    public function queue($view, $queue = null);
 
     /**
      * Queue a new e-mail message for sending after (n) seconds.
      *
-     * @param  int  $delay
-     * @param  string|array  $view
-     * @param  array  $data
-     * @param  \Closure|string  $callback
+     * @param  \DateTime|int  $delay
+     * @param  string|array|MailableContract  $view
      * @param  string  $queue
      * @return mixed
      */
-    public function later($delay, $view, array $data, $callback, $queue = null);
+    public function later($delay, $view, $queue = null);
 }
