@@ -16,7 +16,7 @@ interface Mailer
     /**
      * Send a new message using a view.
      *
-     * @param  string|array  $view
+     * @param  string|array|MailableContract  $view
      * @param  array  $data
      * @param  \Closure|string  $callback
      * @return void
