@@ -29,7 +29,7 @@ trait InteractsWithPivotTable
         // checking which of the given ID/records is in the list of current records
         // and removing all of those rows from this "intermediate" joining table.
         $detach = array_values(array_intersect(
-            $this->newPivotQuery()->pluck($this->relatedKey)->all(),
+            $this->newPivotQuery()->pluck($this->relatedPivotKey)->all(),
             array_keys($records)
         ));
 
@@ -89,7 +89,7 @@ trait InteractsWithPivotTable
         // in this joining table. We'll spin through the given IDs, checking to see
         // if they exist in the array of current ones, and if not we will insert.
         $current = $this->newPivotQuery()->pluck(
-            $this->relatedKey
+            $this->relatedPivotKey
         )->all();
 
         $detach = array_diff($current, array_keys(
@@ -287,9 +287,9 @@ trait InteractsWithPivotTable
      */
     protected function baseAttachRecord($id, $timed)
     {
-        $record[$this->relatedKey] = $id;
+        $record[$this->relatedPivotKey] = $id;
 
-        $record[$this->foreignKey] = $this->parent->getKey();
+        $record[$this->foreignPivotKey] = $this->parent->getKey();
 
         // If the record needs to have creation and update timestamps, we will make
         // them by calling the parent model's "freshTimestamp" method which will
@@ -353,7 +353,7 @@ trait InteractsWithPivotTable
                 return 0;
             }
 
-            $query->whereIn($this->relatedKey, (array) $ids);
+            $query->whereIn($this->relatedPivotKey, (array) $ids);
         }
 
         // Once we have all of the conditions set on the statement, we are ready
@@ -381,7 +381,7 @@ trait InteractsWithPivotTable
             $this->parent, $attributes, $this->table, $exists, $this->using
         );
 
-        return $pivot->setPivotKeys($this->foreignKey, $this->relatedKey);
+        return $pivot->setPivotKeys($this->foreignPivotKey, $this->relatedPivotKey);
     }
 
     /**
@@ -413,7 +413,7 @@ trait InteractsWithPivotTable
      */
     public function newPivotStatementForId($id)
     {
-        return $this->newPivotQuery()->where($this->relatedKey, $id);
+        return $this->newPivotQuery()->where($this->relatedPivotKey, $id);
     }
 
     /**
@@ -433,7 +433,7 @@ trait InteractsWithPivotTable
             call_user_func_array([$query, 'whereIn'], $arguments);
         }
 
-        return $query->where($this->foreignKey, $this->parent->getKey());
+        return $query->where($this->foreignPivotKey, $this->parent->getKey());
     }
 
     /**
