@@ -169,11 +169,16 @@ class Handler implements ExceptionHandlerContract
      */
     protected function prepareResponse($request, Exception $e)
     {
-        if ($this->isHttpException($e)) {
-            return $this->toIlluminateResponse($this->renderHttpException($e), $e);
-        } else {
+        if (! $this->isHttpException($e) && config('app.debug')) {
             return $this->toIlluminateResponse($this->convertExceptionToResponse($e), $e);
         }
+
+        $e = $this->isHttpException($e)
+                ? $e : new HttpException(500, $e->getMessage());
+
+        return $this->toIlluminateResponse(
+            $this->renderHttpException($e), $e
+        );
     }
 
     /**
