@@ -18,19 +18,21 @@ class CookieTest extends TestCase
     {
         $cookie = $this->getCreator();
         $cookie->setDefaultPathAndDomain('foo', 'bar');
-        $c = $cookie->make('color', 'blue', 10, '/path', '/domain', true, false);
+        $c = $cookie->make('color', 'blue', 10, '/path', '/domain', true, false, false, 'lax');
         $this->assertEquals('blue', $c->getValue());
         $this->assertFalse($c->isHttpOnly());
         $this->assertTrue($c->isSecure());
         $this->assertEquals('/domain', $c->getDomain());
         $this->assertEquals('/path', $c->getPath());
+        $this->assertEquals('lax', $c->getSameSite());
 
-        $c2 = $cookie->forever('color', 'blue', '/path', '/domain', true, false);
+        $c2 = $cookie->forever('color', 'blue', '/path', '/domain', true, false, false, 'strict');
         $this->assertEquals('blue', $c2->getValue());
         $this->assertFalse($c2->isHttpOnly());
         $this->assertTrue($c2->isSecure());
         $this->assertEquals('/domain', $c2->getDomain());
         $this->assertEquals('/path', $c2->getPath());
+        $this->assertEquals('strict', $c2->getSameSite());
 
         $c3 = $cookie->forget('color');
         $this->assertNull($c3->getValue());
@@ -40,13 +42,13 @@ class CookieTest extends TestCase
     public function testCookiesAreCreatedWithProperOptionsUsingDefaultPathAndDomain()
     {
         $cookie = $this->getCreator();
-        $cookie->setDefaultPathAndDomain('/path', '/domain');
-        $c = $cookie->make('color', 'blue', 10, null, null, true, false);
+        $cookie->setDefaultPathAndDomain('/path', '/domain', true, 'lax');
+        $c = $cookie->make('color', 'blue');
         $this->assertEquals('blue', $c->getValue());
-        $this->assertFalse($c->isHttpOnly());
         $this->assertTrue($c->isSecure());
         $this->assertEquals('/domain', $c->getDomain());
         $this->assertEquals('/path', $c->getPath());
+        $this->assertEquals('lax', $c->getSameSite());
     }
 
     public function testQueuedCookies()
