@@ -3,10 +3,12 @@
 namespace Illuminate\Tests\Http;
 
 use Mockery as m;
+use JsonSerializable;
 use Illuminate\Http\Request;
 use PHPUnit\Framework\TestCase;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Contracts\Support\Jsonable;
+use Illuminate\Contracts\Support\Arrayable;
 
 class HttpResponseTest extends TestCase
 {
@@ -17,8 +19,16 @@ class HttpResponseTest extends TestCase
 
     public function testJsonResponsesAreConvertedAndHeadersAreSet()
     {
+        $response = new \Illuminate\Http\Response(new ArrayableStub);
+        $this->assertEquals('{"foo":"bar"}', $response->getContent());
+        $this->assertEquals('application/json', $response->headers->get('Content-Type'));
+
         $response = new \Illuminate\Http\Response(new JsonableStub);
         $this->assertEquals('foo', $response->getContent());
+        $this->assertEquals('application/json', $response->headers->get('Content-Type'));
+
+        $response = new \Illuminate\Http\Response(new ArrayableAndJsonableStub);
+        $this->assertEquals('{"foo":"bar"}', $response->getContent());
         $this->assertEquals('application/json', $response->headers->get('Content-Type'));
 
         $response = new \Illuminate\Http\Response();
@@ -29,6 +39,17 @@ class HttpResponseTest extends TestCase
         $response = new \Illuminate\Http\Response(new JsonSerializableStub);
         $this->assertEquals('{"foo":"bar"}', $response->getContent());
         $this->assertEquals('application/json', $response->headers->get('Content-Type'));
+    }
+
+    public function testResponseHeaderTypeIsReset()
+    {
+        $response = new \Illuminate\Http\Response(new ArrayableStub);
+        $this->assertEquals('{"foo":"bar"}', $response->getContent());
+        $this->assertEquals('application/json', $response->headers->get('Content-Type'));
+
+        $response->setContent('foo');
+        $this->assertEquals('foo', $response->getContent());
+        $this->assertNotEquals('application/json', $response->headers->get('Content-Type'));
     }
 
     public function testRenderablesAreRendered()
@@ -132,6 +153,26 @@ class HttpResponseTest extends TestCase
         $response->withErrors($provider);
     }
 
+    public function testWithHeaders()
+    {
+        $response = new \Illuminate\Http\Response(null, 200, ['foo' => 'bar']);
+        $this->assertSame('bar', $response->headers->get('foo'));
+
+        $response->withHeaders(['foo' => 'BAR', 'bar' => 'baz']);
+        $this->assertSame('BAR', $response->headers->get('foo'));
+        $this->assertSame('baz', $response->headers->get('bar'));
+
+        $responseMessageBag = new \Symfony\Component\HttpFoundation\ResponseHeaderBag(['bar' => 'BAZ', 'titi' => 'toto']);
+        $response->withHeaders($responseMessageBag);
+        $this->assertSame('BAZ', $response->headers->get('bar'));
+        $this->assertSame('toto', $response->headers->get('titi'));
+
+        $headerBag = new \Symfony\Component\HttpFoundation\HeaderBag(['bar' => 'BAAA', 'titi' => 'TATA']);
+        $response->withHeaders($headerBag);
+        $this->assertSame('BAAA', $response->headers->get('bar'));
+        $this->assertSame('TATA', $response->headers->get('titi'));
+    }
+
     public function testMagicCall()
     {
         $response = new RedirectResponse('foo.bar');
@@ -151,6 +192,27 @@ class HttpResponseTest extends TestCase
     }
 }
 
+class ArrayableStub implements Arrayable
+{
+    public function toArray()
+    {
+        return ['foo' => 'bar'];
+    }
+}
+
+class ArrayableAndJsonableStub implements Arrayable, Jsonable
+{
+    public function toJson($options = 0)
+    {
+        return '{"foo":"bar"}';
+    }
+
+    public function toArray()
+    {
+        return [];
+    }
+}
+
 class JsonableStub implements Jsonable
 {
     public function toJson($options = 0)
@@ -159,7 +221,7 @@ class JsonableStub implements Jsonable
     }
 }
 
-class JsonSerializableStub implements \JsonSerializable
+class JsonSerializableStub implements JsonSerializable
 {
     public function jsonSerialize()
     {
